Fix UT.
One incorrect test case is removed. This test case is totally wrong.
The spinner is for image uploading to EPS, so it should be shown after user selects the file and before the upload is done. For this case, it should be shown after input's change event and before loadData is done. That means, after loadData is finished, the spinner should be hidden because the upload is done (pay attention that this spinner is NOT the one for posting ad).
The previous code uses toggleClass which is against best practice because it will have wrong result when previous status is wrong. As input change event is not triggered by code, the "initial" status is incorrect, that makes the result consistent with incorrect expectation.
As we mocks the upload progress, it's very hard to check the status during the upload in current mechanism. So this case will be removed instead of reimplemented.