Fix the problem that post 2.0 page will be accessed twice in production environment
The post 2.0 page is accessed twice because it enables ABTest. However, if user is in post 2.0 page, it means he already is selected to use 2.0 in the ABTesting. So we don't need to do that in post page.

PS: There is no way to pass the cookie with ABTest because the user redirect is done completely in GA and node server has no clue where the user will be redirect to.
PPS: Post 2.0 page will never be the ABTest page because if new user comes in, it will be redirected to 1.0 post page without running any ABTesting code in server or client.

PPS: The following conclusion seems so weird but that's what I get from code. It has nothing to do with current fix. But I put it here in case someone asks about it and thinks it's a new introduced regression defect.
If a new user without any cookie goes to MX site, he will never see 2.0 post.
1. When he accesses homepage, as we have turned off ABTesting in homepage, there will be no set in cookie, so his cookie reveals he is using 1.0
2. When he access post page via /post, the site finds it's not a 2.0 user, so it will redirect him to /post.html
3. There is no ABTesting in 1.0 post, so he will always be in 1.0 until he uses search page or VIP page which has enabled ABTesting.