Move the logic to remove cookie from server side to client side.
The cookies are designed to be used only once. However, ABTest prevents us from removing cookie in server side. The removal of cookie has to be executed after ABTest redirect so it should be executed as late as possible. This change move the logic to client and make it be the last code run on page load.